Completed assigntment

+++ thirdAssignment/Controllers/PatientController.cs
@@ -4,11 +4,11 @@ using Microsoft.IdentityModel.Tokens;
 using thirdAssignment.Aplication.Core;
 using thirdAssignment.Aplication.Dtos;
 using thirdAssignment.Aplication.Interfaces.Contracts;
-using thirdAssignment.Aplication.Models;
+using thirdAssignment.Aplication.Models.Doctor;
+using thirdAssignment.Aplication.Models.Patient;
 using thirdAssignment.Aplication.Services;
 using thirdAssignment.Presentation.Models;
 using thirdAssignment.Presentation.Utils;
-using thirdAssignment.Presentation.Utils.SessionHandler;
 using thirdAssignment.Presentation.Utils.UserValidations;
 
 namespace thirdAssignment.Presentation.Controllers
@@ -29,13 +29,14 @@ namespace thirdAssignment.Presentation.Controllers
         public async Task<IActionResult> Index()
         {
             if (!_userValidations.HasUser()) return RedirectToAction("Login", "User");
+            if (!_userValidations.UserIsAssistent()) return RedirectToAction("index", "Home");
 
-            Result<List<PatientModel>> result = new();
+            Result<List<Aplication.Models.Patient.PatientModel>> result = new();
             try
             {
-                var currentUser = HttpContext.Session.Get<UserModel>("user");
 
-                result = await _patientService.GetAll(currentUser.ConsultingRoom.Id);
+
+                result = await _patientService.GetAll();
 
                 if (!result.IsSuccess)
                 {
@@ -59,6 +60,8 @@ namespace thirdAssignment.Presentation.Controllers
         public async Task<IActionResult> SavePatient()
         {
 			if (!_userValidations.HasUser()) return RedirectToAction("Login", "User");
+            if (!_userValidations.UserIsAssistent()) return RedirectToAction("index", "Home");
+
             ViewBag.IsSmoker= true;
             ViewBag.HasAllergies = true;
 
@@ -68,26 +71,24 @@ namespace thirdAssignment.Presentation.Controllers
         // POST: PatientController/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> SavePatient(SavePatientDto saveDto, int smokes, int hasAlergies)
+        public async Task<IActionResult> SavePatient(PatienSaveModel saveDto, int smokes, int hasAlergies)
         
         {
             //IFormFile img
             if (!_userValidations.HasUser()) return RedirectToAction("Login", "User");
+            if (!_userValidations.UserIsAssistent()) return RedirectToAction("index", "Home");
 
-            Result<PatientModel> result = new();
+            Result<PatienSaveModel> result = new();
             try
             {
+
 
                 if (!ModelState.IsValid)
                 {
                     ViewBag.message = ModelState.Values.SelectMany(v => v.Errors).First().ErrorMessage;
                     return View(_generateSelectList.GenereteCheckBoxList());
                 }
-                if (saveDto.ImgPath.IsNullOrEmpty())
-                {
-                    ViewBag.message = "the Imge field is requierd";
-                    return View();
-                }
+
 
                 //if (img != null)
                 //{
@@ -104,34 +105,91 @@ namespace thirdAssignment.Presentation.Controllers
 
                 saveDto.HasAllergies = hasAlergies == 1? true : false;
 
-                saveDto.ConsultingRoomId = HttpContext.Session.Get<UserModel>("user").ConsultingRoom.Id;
-
                 result = await _patientService.Save(saveDto);
+
+
+
+
+                PatienSaveModel savedDoctor = result.Data;
+
+                if (savedDoctor != null && savedDoctor.Id != null)
+                {
+                    savedDoctor.ImgPath = UploadFile(saveDto.File, savedDoctor.Id);
+
+                    result = await _patientService.Update(savedDoctor);
+                }
+
+
 
                 if (!result.IsSuccess)
                 {
                     ViewBag.Message = result.Message;
-                    return View();
+                    return View(_generateSelectList.GenereteCheckBoxList());
                 }
 
                 return RedirectToAction(nameof(Index));
             }
             catch
             {
-                return View();
+                return View(_generateSelectList.GenereteCheckBoxList());
             }
+        }
+
+        private string UploadFile(IFormFile filePath, Guid id, bool editmode = false, string ImgUrl = "")
+        {
+            if (editmode && filePath == null)
+            {
+                return ImgUrl;
+            }
+            //Get directory path
+            string basePath = $"/Images/Patients/{id}";
+            string path = Path.Combine(Directory.GetCurrentDirectory(), $"wwwroot{basePath}");
+
+            //create folder if not exits
+            if (!Directory.Exists(path))
+            {
+                Directory.CreateDirectory(path);
+            }
+
+            // get file path
+            Guid guid = Guid.NewGuid();
+
+            FileInfo fileInfo = new(filePath.FileName);
+
+            string filename = guid + fileInfo.Extension;
+
+            string fileNameWithPath = Path.Combine(path, filename);
+
+            using (var stream = new FileStream(fileNameWithPath, FileMode.Create))
+            {
+                filePath.CopyTo(stream);
+            }
+
+            if (editmode)
+            {
+                string[] oldImgUrlPort = ImgUrl.Split("/");
+                string oldImageName = oldImgUrlPort[^1];
+                string CompleteOldPath = Path.Combine(path, oldImageName);
+                if (System.IO.File.Exists(CompleteOldPath))
+                {
+                    System.IO.File.Delete(CompleteOldPath); 
+                }
+			}
+            return $"{basePath}/{ filename}";
         }
 
         // GET: PatientController/Edit/5
         public async Task<IActionResult> EditPatient(Guid id)
         {
             if (!_userValidations.HasUser()) return RedirectToAction("Login", "User");
+            if (!_userValidations.UserIsAssistent()) return RedirectToAction("index", "Home");
 
-            Result<PatientModel> result = new();
+            Result<Aplication.Models.Patient.PatientModel> result = new();
             try
             {
-
-                result = await _patientService.GetById(id);
+				
+			
+				result = await _patientService.GetById(id);
 
                 if (!result.IsSuccess)
                 {
@@ -152,32 +210,41 @@ namespace thirdAssignment.Presentation.Controllers
         // POST: PatientController/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> EditPatient(Guid id, UpdatePatientDto updateDto, int smokes, int hasAlergies)
+        public async Task<IActionResult> EditPatient(Guid id, PatienSaveModel updateDto, int smokes, int hasAlergies)
         {
             if (!_userValidations.HasUser()) return RedirectToAction("Login", "User");
+            if (!_userValidations.UserIsAssistent()) return RedirectToAction("index", "Home");
 
-            Result<PatientModel> result = new();
+            Result<PatienSaveModel> result = new();
             try
             {
 
 
                 if (!ModelState.IsValid)
                 {
-                    result = await _patientService.GetById(id);
-                   ViewBag.message = ModelState.Values.SelectMany(v => v.Errors).First().ErrorMessage;
-                    return View(new EditPatientModel { Checkboxes = _generateSelectList.GenereteCheckBoxList(result.Data), patientModel = result.Data });
+                    Result<Aplication.Models.Patient.PatientModel> resultIner = new();
+
+                    resultIner = await _patientService.GetById(id);
+                    ViewBag.message = ModelState.Values.SelectMany(v => v.Errors).First().ErrorMessage;
+                    return View(new EditPatientModel { Checkboxes = _generateSelectList.GenereteCheckBoxList(resultIner.Data), patientModel = resultIner.Data });
                 }
 
                 updateDto.IsSmoker = smokes == 1 ? true : false;
 
                 updateDto.HasAllergies = hasAlergies == 1 ? true : false;
 
-                result = await _patientService.Update(updateDto);
+                PatientModel patient = _patientService.GetById(updateDto.Id).Result.Data;
+				
+                    
+                updateDto.ImgPath = UploadFile(updateDto.File, patient.Id, true, patient.ImgPath);
+
+				result = await _patientService.Update(updateDto);
+
 
 
                 if (!result.IsSuccess)
                 {
-                    Result<PatientModel> resultIner = new();
+                    Result<Aplication.Models.Patient.PatientModel> resultIner = new();
 
                     resultIner = await _patientService.GetById(id);
                     ViewBag.Message = result.Message;
@@ -195,8 +262,9 @@ namespace thirdAssignment.Presentation.Controllers
         public async Task<IActionResult> DeletePatient(Guid id)
         {
             if (!_userValidations.HasUser()) return RedirectToAction("Login", "User");
+            if (!_userValidations.UserIsAssistent()) return RedirectToAction("index", "Home");
 
-            Result<PatientModel> result = new();
+            Result<Aplication.Models.Patient.PatientModel> result = new();
             try
             {
                 result = await _patientService.GetById(id);
@@ -207,9 +275,9 @@ namespace thirdAssignment.Presentation.Controllers
                     return RedirectToAction(nameof(Index));
                 }
 
-                return View(result.Data);
+				return View(result.Data);
 
-            }
+			}
             catch
             {
 
@@ -223,8 +291,9 @@ namespace thirdAssignment.Presentation.Controllers
         public async Task<IActionResult> DeletePatient(Guid id, IFormCollection collection)
         {
             if (!_userValidations.HasUser()) return RedirectToAction("Login", "User");
+            if (!_userValidations.UserIsAssistent()) return RedirectToAction("index", "Home");
 
-            Result<PatientModel> result = new();
+            Result<Aplication.Models.Patient.PatientModel> result = new();
             try
             {
                 result = await _patientService.Delete(id);
@@ -234,7 +303,27 @@ namespace thirdAssignment.Presentation.Controllers
                     ViewBag.Message = result.Message;
                     return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+
+				string basePath = $"/Images/Patients/{id}";
+				string path = Path.Combine(Directory.GetCurrentDirectory(), $"wwwroot{basePath}");
+				if (Directory.Exists(path))
+				{
+					DirectoryInfo directoryInfo = new DirectoryInfo(path);
+					foreach (FileInfo file in directoryInfo.GetFiles())
+					{
+						file.Delete();
+					}
+
+					foreach (DirectoryInfo folder in directoryInfo.GetDirectories())
+					{
+						folder.Delete(true);
+					}
+				}
+			
+
+
+
+				return RedirectToAction(nameof(Index));
 
             }
             catch
